SCRD-5089 openstack-ardana: improved Jenkins artifacts collection
Improves the collection of Jenkins artifacts by passing this
responsibility to the master openstack-ardana job, if one is
present. The tempest JUnit test case results are also collected
by the master openstack-ardana job, which allows a tracking record
to be maintained independently by each main job.

+++ jenkins/ci.suse.de/pipelines/openstack-ardana-bootstrap-clm.Jenkinsfile
@@ -73,10 +73,15 @@ pipeline {
 
   post {
     always {
-        // archiveArtifacts doesn't support absolute paths, so we have to to this instead
-        sh 'ln -s ${SHARED_WORKSPACE} ${BUILD_NUMBER}'
-        archiveArtifacts artifacts: "${BUILD_NUMBER}/.artifacts/**/*", allowEmptyArchive: true
-        sh 'rm ${BUILD_NUMBER}'
+      script {
+        // Let the upstream job archive artifacts
+        if (reuse_node == '') {
+          // archiveArtifacts doesn't support absolute paths, so we have to to this instead
+          sh 'ln -s ${SHARED_WORKSPACE} ${BUILD_NUMBER}'
+          archiveArtifacts artifacts: "${BUILD_NUMBER}/.artifacts/**/*", allowEmptyArchive: true
+          sh 'rm ${BUILD_NUMBER}'
+        }
+      }
     }
   }
 }
